+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 - Nodepool nodes are labeled with nodepool id on AWS using `giantswarm.io/machine-pool`.
 - `MachinePool` and `AzureMachinePool` are labeled with the `giantswarm.io/machine-pool` label.
+- `get releases` command to return details of available releases.
 
 ## [1.40.0] - 2021-09-24